refactor: 예약 생성시 검증 추가
- 기존 dto를 제거하고 검증 로직이 포함된 vo 구현
- 무결성을 보장하기 위해 Reservation에도 검증 로직 추가
