--- src/main/java/roomescape/dto/ReservationCreateDto.java
@@ -1,7 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record ReservationCreateDto(String name, LocalDate date, LocalTime time) {
-}
