+++ src/main/java/roomescape/vo/ReservationCreateVo.java
@@ -1,24 +1,18 @@
-package roomescape.model;
+package roomescape.vo;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class Reservation {
-    private Long id;
-    private String name;
-    private LocalDate date;
-    private LocalTime time;
+public class ReservationCreateVo {
+    private final String name;
+    private final LocalDate date;
+    private final LocalTime time;
 
-    public Reservation(Long id, String name, LocalDate date, LocalTime time) {
+    public ReservationCreateVo(String name, LocalDate date, LocalTime time) {
         validate(name, date, time);
-        this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
-    }
-
-    public long getId() {
-        return id;
     }
 
     public String getName() {
